feat: Multiagent with different vehicle types I
