+++ ardupilot_gz_bringup/launch/robots/wildthumper.launch.py
@@ -34,134 +34,15 @@ master:=tcp:127.0.0.1:5760
 sitl:=127.0.0.1:5501
 """
 import os
+import tempfile
 
 from ament_index_python.packages import get_package_share_directory
 from launch import LaunchDescription
-from launch.actions import (DeclareLaunchArgument, IncludeLaunchDescription,
-                            OpaqueFunction, RegisterEventHandler)
-from launch.conditions import IfCondition
-from launch.event_handlers import OnProcessStart
+from launch.actions import DeclareLaunchArgument, IncludeLaunchDescription
 from launch.launch_description_sources import PythonLaunchDescriptionSource
 from launch.substitutions import LaunchConfiguration, PathJoinSubstitution
-from launch_ros.actions import Node
 from launch_ros.substitutions import FindPackageShare
 
-
-def launch_spawn_robot(context):
-    """Return a Gazebo spawn robot launch description"""
-    # Get substitutions for arguments
-    name = LaunchConfiguration("name")
-    pos_x = LaunchConfiguration("x")
-    pos_y = LaunchConfiguration("y")
-    pos_z = LaunchConfiguration("z")
-    rot_r = LaunchConfiguration("R")
-    rot_p = LaunchConfiguration("P")
-    rot_y = LaunchConfiguration("Y")
-
-    # spawn robot
-    spawn_robot =  Node(
-        package="ros_gz_sim",
-        executable="create",
-        namespace=name,
-        arguments=[
-            "-world",
-            "",
-            "-param",
-            "",
-            "-name",
-            name,
-            "-topic",
-            "/robot_description",
-            "-x",
-            pos_x,
-            "-y",
-            pos_y,
-            "-z",
-            pos_z,
-            "-R",
-            rot_r,
-            "-P",
-            rot_p,
-            "-Y",
-            rot_y,
-        ],
-        output="screen",
-    )
-    return [spawn_robot]
-
-
-def launch_state_pub_with_bridge(context):
-    model_name = LaunchConfiguration("model").perform(context)
-    pkg_ardupilot_sitl_models = get_package_share_directory("ardupilot_sitl_models")
-    pkg_project_bringup = get_package_share_directory("ardupilot_gz_bringup")
-    
-    # Load SDF file.
-    sdf_file = os.path.join(
-        pkg_ardupilot_sitl_models, "models", model_name, "model.sdf"
-    )
-    with open(sdf_file, "r") as infp:
-        robot_desc = infp.read()
-
-        # substitute `models://` with `package://ardupilot_sitl_models/models/`
-        # for sdformat_urdf plugin used by robot_state_publisher
-        robot_desc = robot_desc.replace(
-            "model://wildthumper",
-            "package://ardupilot_sitl_models/models/wildthumper")
-
-        robot_desc = robot_desc.replace(
-            "model://wildthumper_with_lidar",
-            "package://ardupilot_sitl_models/models/wildthumper_with_lidar")
-
-    # Publish /tf and /tf_static.
-    robot_state_publisher = Node(
-        package="robot_state_publisher",
-        executable="robot_state_publisher",
-        name="robot_state_publisher",
-        output="both",
-        parameters=[
-            {"robot_description": robot_desc},
-            {"frame_prefix": ""},
-        ],
-    )
-
-    # Bridge.
-    bridge = Node(
-        package="ros_gz_bridge",
-        executable="parameter_bridge",
-        parameters=[
-            {
-                "config_file": os.path.join(
-                    pkg_project_bringup, "config", "wildthumper_bridge.yaml"
-                ),
-                "qos_overrides./tf_static.publisher.durability": "transient_local",
-            }
-        ],
-        output="screen",
-    )
-
-    # Relay - use instead of transform when Gazebo is only publishing odom -> base_link
-    topic_tools_tf = Node(
-        package="topic_tools",
-        executable="relay",
-        arguments=[
-            "/gz/tf",
-            "/tf",
-        ],
-        output="screen",
-        respawn=False,
-        condition=IfCondition(LaunchConfiguration("use_gz_tf")),
-    )
-    
-    event = RegisterEventHandler(
-        OnProcessStart(
-            target_action=bridge,
-            on_start=[
-                topic_tools_tf
-            ]
-        )
-    )
-    
-    return [robot_state_publisher, bridge, event]
 
 def generate_launch_arguments():
     """Generate a list of launch arguments"""
@@ -172,11 +53,6 @@ def generate_launch_arguments():
             description="Use Gazebo TF."
         ),
         # Gazebo model launch arguments.
-        DeclareLaunchArgument(
-            "model",
-            default_value="wildthumper_with_lidar",
-            description="Name or filepath of the model to load.",
-        ),
         DeclareLaunchArgument(
             "name",
             default_value="wildthumper",
@@ -212,6 +88,17 @@ def generate_launch_arguments():
             default_value="0",
             description="The intial yaw angle (radians).",
         ),
+        DeclareLaunchArgument(
+            "instance",
+            default_value="0",
+            description="Set instance of SITL "
+            "(adds 10*instance to all port numbers).",
+        ),
+        DeclareLaunchArgument(
+            "sysid",
+            default_value="",
+            description="Set SYSID_THISMAV.",
+        ),
     ]
 
 
@@ -220,66 +107,73 @@ def generate_launch_description():
     
     launch_arguments = generate_launch_arguments()
     
+    pkg_ardupilot_sitl_models = get_package_share_directory("ardupilot_sitl_models")
+    pkg_project_bringup = get_package_share_directory("ardupilot_gz_bringup")
     pkg_ardupilot_sitl = get_package_share_directory("ardupilot_sitl")
+    
+    # Load SDF file.
+    sdf_file = os.path.join(
+        pkg_ardupilot_sitl_models, "models", "wildthumper_with_lidar", "model.sdf"
+    )
+    with open(sdf_file, "r") as infp:
+        robot_desc = infp.read()
 
-    # Include component launch files.
-    sitl_dds = IncludeLaunchDescription(
+        # substitute `models://` with `package://ardupilot_sitl_models/models/`
+        # for sdformat_urdf plugin used by robot_state_publisher
+        robot_desc = robot_desc.replace(
+            "model://wildthumper",
+            "package://ardupilot_sitl_models/models/wildthumper")
+
+        robot_desc = robot_desc.replace(
+            "model://wildthumper_with_lidar",
+            "package://ardupilot_sitl_models/models/wildthumper_with_lidar")
+
+    temp_file = tempfile.NamedTemporaryFile(delete=False, suffix=".yaml")
+    sdf_file_modified = temp_file.name
+
+    with open(sdf_file_modified, 'w') as temp_file:
+        temp_file.write(robot_desc)
+    
+    sitl_config_file = os.path.join(
+        pkg_ardupilot_sitl, "config", "default_params", "rover-skid.parm"
+    )
+
+    bridge_config_file = os.path.join(
+        pkg_project_bringup, "config", "wildthumper_bridge.yaml"
+    )
+    
+    robot = IncludeLaunchDescription(
         PythonLaunchDescriptionSource(
             [
                 PathJoinSubstitution(
                     [
-                        FindPackageShare("ardupilot_sitl"),
+                        FindPackageShare("ardupilot_gz_bringup"),
                         "launch",
-                        "sitl_dds_udp.launch.py",
+                        "robots",
+                        "robot.launch.py",
                     ]
                 ),
             ]
         ),
         launch_arguments={
-            "transport": "udp4",
-            "port": "2019",
-            "command": "ardurover",
-            "synthetic_clock": "True",
-            "wipe": "False",
-            "model": "json",
-            "speedup": "1",
-            "slave": "0",
-            "instance": "0",
-            "defaults": os.path.join(
-                pkg_ardupilot_sitl,
-                "config",
-                "default_params",
-                "rover-skid.parm",
-            )
-            + ","
-            + os.path.join(
-                pkg_ardupilot_sitl,
-                "config",
-                "default_params",
-                "dds_udp.parm",
-            ),
-            "sim_address": "127.0.0.1",
-            "master": "tcp:127.0.0.1:5760",
-            "sitl": "127.0.0.1:5501",
+            "use_gz_tf": LaunchConfiguration("use_gz_tf"),
+            "sdf_file": sdf_file_modified,
+            "sitl_config_file": sitl_config_file,
+            "bridge_config_file": bridge_config_file,
+            "name": LaunchConfiguration("name"),
+            "x": LaunchConfiguration("x"),
+            "y": LaunchConfiguration("y"),
+            "z": LaunchConfiguration("z"),
+            "R": LaunchConfiguration("R"),
+            "P": LaunchConfiguration("P"),
+            "Y": LaunchConfiguration("Y"),
+            "instance": LaunchConfiguration("instance"),
+            "sysid": LaunchConfiguration("sysid"),
         }.items(),
     )
 
-    # Ensure `SDF_PATH` is populated as `sdformat_urdf`` uses this rather
-    # than `GZ_SIM_RESOURCE_PATH` to locate resources.
-    if "GZ_SIM_RESOURCE_PATH" in os.environ:
-        gz_sim_resource_path = os.environ["GZ_SIM_RESOURCE_PATH"]
-
-        if "SDF_PATH" in os.environ:
-            sdf_path = os.environ["SDF_PATH"]
-            os.environ["SDF_PATH"] = sdf_path + ":" + gz_sim_resource_path
-        else:
-            os.environ["SDF_PATH"] = gz_sim_resource_path
-
-    opfunc_robot_state_publisher = OpaqueFunction(function=launch_state_pub_with_bridge)
-    opfunc_spawn_robot = OpaqueFunction(function=launch_spawn_robot)
-    ld = LaunchDescription(launch_arguments)
-    ld.add_action(opfunc_robot_state_publisher)
-    ld.add_action(sitl_dds)
-    ld.add_action(opfunc_spawn_robot)
-
-    return ld
+    return LaunchDescription(
+        launch_arguments + [
+            robot,
+        ]
+    )
