+++ ardupilot_gz_bringup/launch/wildthumper_playpen.launch.py
@@ -63,7 +63,6 @@ def generate_launch_description():
             ]
         ),
         launch_arguments={
-            "model": "wildthumper_with_lidar",
             "name": "wildthumper",
             "x": "0.0",
             "y": "0.0",
@@ -95,6 +94,7 @@ def generate_launch_description():
     # RViz.
     rviz = Node(
         package="rviz2",
+        namespace="wildthumper",
         executable="rviz2",
         arguments=["-d", f'{Path(pkg_project_bringup) / "rviz" / "wildthumper.rviz"}'],
         condition=IfCondition(LaunchConfiguration("rviz")),
